move AQI_STANDARDS to variables
