+++ components/AQIStandard/index.js
@@ -1,6 +1,6 @@
 import clsx from "clsx";
 import React from "react";
-import { AQI_STANDARDS } from "./data";
+import { AQI_STANDARDS } from "../../utils/variables";
 
 function AQIStandard({ className, type }) {
   const { label, bgColor, color } = AQI_STANDARDS[type];
